Big-Integer- Arithmetic
Developed a program that implements arithmetic with large integers, of
arbitrary size. Numbers were represented as linked list of integers,
where the digits are in higher base to reduce the processing time.
Implemented the arithmetic for negative long integers as well.
Written methods to implement all basic arithmetic operations on big
numbers which ensured minimal execution time.
Implemented methods are as follows:

1. BigNum(String s): Constructor for BigNum class; takes a string s as
parameter, that stores a number in decimal, and creates the BigNum
object representing that number. The string can have arbitrary length.

2. BigNum(Long num): Constructor for BigNum class.

3. String toString(): convert the BigNum class object into its
equivalent string (in decimal). There should be no leading zeroes in the
string.

4. BigNum add(BigNum a, BigNum b): sum of two numbers stored as BigNum.

5. BigNum subtract(BigNum a, BigNum b): given two BigNum a and b as
parameters, representing the numbers n1 and n2 repectively, returns the
BigNum corresponding to n1-n2.

5. BigNum product(BigNum a, BigNum b): product of two numbers.

6. BigNum power(BigNum a, long n): given an BigNum a, representing the
number x and n, returns the BigNum corresponding to x^n (x to the power
n).

7. printList(): Print the base + ":" + elements of the list, separated
by spaces.

8. BigNum power(BigNum a, BigNum n): return a^n, where a and n are both
BigNum. Here a may be negative, but assumed that n is non-negative.

9. BigNum divide(BigNum a, BigNum b): Divide a by b result. Fractional
part is discarded (take just the quotient). Both a and b may be positive
or negative. If b is 0, program raises an exception.

10. BigNum mod(BigNum a, BigNum b): remainder you get when a is divided
by b (a%b). Assumed that a is non-negative, and b > 0.

11. BigNum squareRoot(BigNum a): return the square root of a
(truncated). Used binary search. Assumed that a is non-negative.