Shortest-Path-Algorithm
Implemented the following shortest path algorithms.
1. BFS
2. Dijkstra's algorithm
3. DAG shortest paths
4. Bellman-Ford algorithm

Wrote a program that behaves as follows:
given a directed graph as input, if the graph has uniform weights (i.e.,
same positive weights for all edges), then it runs BFS to find shorest
paths. Otherwise, if the graph is a directed, acyclic graph (DAG), then
it runs DAG shortest paths. Otherwise, if the graph has only non
negative weights, then it runs Dijkstra's algorithm. If all these test
fail, then it runs the Bellman-Ford algorithm. If the graph has negative
cycles, then it prints the message "Unable to solve problem. Graph has a
negative cycle".

Also implemented the additional functionalities as follows.
The number of paths in a DAG from source s to each vertex u in G can be
easily calculated in linear time. If G does not contain a negative or
zero cycle reachable from s, then it can be shown that the set of all
shortest paths from s forms a DAG (by selecting only those edges e=(u,v)
with v.distance = u.distance + e.Weight). By using the algorithm for
counting paths in DAGs on this subgraph of G, we can find the number of
shortest paths from s to each node of G. The input is a directed graph
G=(V,E), with edge weights W:E-->Z (negative weights are possible). The
output is the sum of the number of shortest paths (not necessarily
disjoint) from s to each vertex u in the graph. If the graph has a
negative or zero cycle, reachable from s, then print a message
"Non-positive cycle in graph. DAC is not applicable". In addition,
program outputs such a cycle